# openapi.yaml
openapi: 3.1.0
info:
  title: BookClose API
  version: 1.0.0
paths:
  /api/engagements:
    get:
      summary: Get engagements
      tags: [Engagements]
      responses:
        '200':
          description: List of engagements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Engagement'
    post:
      summary: Create an engagement
      tags: [Engagements]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngagementCreate'
      responses:
        '200':
          description: Created engagement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Engagement'
  /api/engagements/{engagement_id}:
    patch:
      summary: Update engagement status
      tags: [Engagements]
      parameters:
        - name: engagement_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        '200':
          description: Updated engagement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Engagement'
...
  /api/engagements/{engagement_id}/sign:
    post:
      summary: Sign engagement letter
      tags: [Engagements]
      parameters:
        - name: engagement_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                signer_id:
                  type: integer
                signature_data:
                  type: string
      responses:
        '200':
          description: Signed engagement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Engagement'
  /api/engagements/{engagement_id}/qbo-sync:
    post:
      summary: Sync engagement with QBO
      tags: [Engagements]
      parameters:
        - name: engagement_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sync status
          content:
            application/json:
              schema:
                type: string
...
  /api/ar/invoices:
    post:
      summary: Create an invoice
      tags: [AR]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                customer_id: { type: integer }
                issue_date: { type: string, format: date-time }
                due_date: { type: string, format: date-time }
                total: { type: number }
                lines: { type: array, items: { type: object } }
      parameters:
        - name: firm_id
          in: query
          required: true
          schema: { type: string }
        - name: client_id
          in: query
          schema: { type: integer }
      responses:
        200: { content: { application/json: { schema: { $ref: '#/components/schemas/Invoice' } } } }
  /api/ar/invoices/{id}:
    get:
      summary: Get invoice details
      tags: [AR]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: firm_id
          in: query
          required: true
          schema: { type: string }
      responses:
        200: { content: { application/json: { schema: { $ref: '#/components/schemas/Invoice' } } } }
    patch:
      summary: Update invoice status
      tags: [AR]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: firm_id
          in: query
          required: true
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        200: { content: { application/json: { schema: { $ref: '#/components/schemas/Invoice' } } } }
  /api/ar/collections/remind:
    post:
      summary: Send reminder for overdue invoice
      tags: [AR]
      parameters:
        - name: invoice_id
          in: query
          required: true
          schema: { type: integer }
        - name: firm_id
          in: query
          required: true
          schema: { type: string }
        - name: client_id
          in: query
          schema: { type: integer }
      responses:
        200: { content: { application/json: { schema: { $ref: '#/components/schemas/Invoice' } } } }
  /api/ar/payments/apply:
    post:
      summary: Apply payment to invoices
      tags: [AR]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreate'
      parameters:
        - name: firm_id
          in: query
          required: true
          schema: { type: string }
        - name: client_id
          in: query
          schema: { type: integer }
      responses:
        200: { content: { application/json: { schema: { $ref: '#/components/schemas/Payment' } } } }
  /api/ar/credits:
    post:
      summary: Create a credit memo
      tags: [AR]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditMemoCreate'
      parameters:
        - name: firm_id
          in: query
          required: true
          schema: { type: string }
        - name: client_id
          in: query
          schema: { type: integer }
      responses:
        200: { content: { application/json: { schema: { $ref: '#/components/schemas/CreditMemo' } } } }
components:
  schemas:
    Engagement:
      type: object
      properties:
        engagement_id:
          type: integer
        entity_id:
          type: integer
        service_ids:
          type: array
          items:
            type: integer
        status:
          type: string
        due_date:
          type: string
          format: date-time
        triage_level:
          type: string
        user_input:
          type: object
        allowance_overage:
          type: number
        agreed_at:
          type: string
          format: date-time
        health_status:
          type: string
        compliance_status:
          type: string
        e_signature:
          type: object
        qbo_sync_status:
          type: string
    EngagementCreate:
      type: object
      properties:
        entity_id:
          type: integer
        service_ids:
          type: array
          items:
            type: integer
        due_date:
          type: string
          format: date-time
        triage_level:
          type: string
        user_input:
          type: object
        allowance_overage:
          type: number
...
    Invoice:
      type: object
      properties:
        invoice_id: { type: integer }
        firm_id: { type: string }
        client_id: { type: integer, nullable: true }
        customer_id: { type: integer }
        qbo_id: { type: string, nullable: true }
        issue_date: { type: string, format: date-time }
        due_date: { type: string, format: date-time }
        total: { type: number }
        lines: { type: array, items: { type: object }, nullable: true }
        status: { type: string }
        attachment_refs: { type: array, items: { type: string }, nullable: true }
    Payment:
      type: object
      properties:
        payment_id: { type: integer }
        firm_id: { type: string }
        client_id: { type: integer, nullable: true }
        qbo_id: { type: string, nullable: true }
        invoice_ids: { type: array, items: { type: integer }, nullable: true }
        amount: { type: number }
        date: { type: string, format: date-time }
        method: { type: string, nullable: true }
    CreditMemo:
      type: object
      properties:
        memo_id: { type: integer }
        firm_id: { type: string }
        client_id: { type: integer, nullable: true }
        qbo_id: { type: string, nullable: true }
        invoice_id: { type: integer }
        amount: { type: number }
        reason: { type: string, nullable: true }